% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_and_factor.R
\name{unnest_and_factor}
\alias{unnest_and_factor}
\title{A Function to Unnest and Factor a Variable in a Tibble}
\usage{
unnest_and_factor(df, var, sep)
}
\arguments{
\item{df}{The name of the stored tibble you want to work with - called 'df' for dataframe based on common stackoverflow usage}

\item{var}{The target variable from that tibble - called 'var' as short for variable}

\item{sep}{The separator string that the function should use to identify nested values; include quotations for proper usage - called 'sep' to mirror dplyr language}
}
\value{
The function returns the unnested tibble.
}
\description{
'unnest_and_factor' takes a designated variable from a designated dataframe and helps to tidy
the data by making sure that if there are multiple values included in a single observation with a standardized
separator, these values are unnested--separated--and then factored for sorting, counting, and other grouping purposes.
}
\examples{
\dontrun{ unnest_and_factor(datateachr::steam_games, game_details,",")
unnest_and_factor(datateachr::apt_buildings, parking_type, ",")
unnest_and_factor(datateachr::apt_buildings, bike_parking, " and ")}

}
\references{
Syntax suggested on StackOverflow by LyzandeR
link: \url{https://stackoverflow.com/questions/48062213/dplyr-using-column-names-as-function-arguments}
More information can aso be found in the book 'Tidy Evaluation' by Lionel Henry and Hadley Wickham
link: \url{https://tidyeval.tidyverse.org/dplyr.html}
}
